# Add NaN/Inf Checks