--- a/geom/alg_convex_hull.go
+++ b/geom/alg_convex_hull.go
@@ -17,7 +17,7 @@ func convexHull(g Geometry) Geometry {
 
 	// Check for point case:
 	if !hasAtLeast2DistinctPointsInXYs(pts) {
-		return pts[0].AsPoint().AsGeometry()
+		return pts[0].asUncheckedPoint().AsGeometry()
 	}
 
 	hull := monotoneChain(pts)