--- a/geom/alg_intersection.go
+++ b/geom/alg_intersection.go
@@ -3,30 +3,36 @@ package geom
 func intersectionOfIndexedLines(
 	lines1 indexedLines, lines2 indexedLines,
 ) (
-	MultiPoint, MultiLineString,
+	MultiPoint, MultiLineString, error,
 ) {
 	// TODO: Investigate potential speed up of swapping lines.
 	var lss []LineString
 	var pts []Point
 	seen := make(map[XY]bool)
 	for i := range lines1.lines {
-		lines2.tree.RangeSearch(lines1.lines[i].envelope().box(), func(j int) error {
+		if err := lines2.tree.RangeSearch(lines1.lines[i].envelope().box(), func(j int) error {
 			inter := lines1.lines[i].intersectLine(lines2.lines[j])
 			if inter.empty {
 				return nil
 			}
 			if inter.ptA == inter.ptB {
-				if pt := inter.ptA; !seen[pt] {
-					pts = append(pts, pt.AsPoint())
-					seen[pt] = true
+				if xy := inter.ptA; !seen[xy] {
+					pt, err := xy.AsPoint()
+					if err != nil {
+						return wrap(err, "intersection between indexed lines is invalid")
+					}
+					pts = append(pts, pt)
+					seen[xy] = true
 				}
 			} else {
 				lss = append(lss, line{inter.ptA, inter.ptB}.asLineString())
 			}
 			return nil
-		})
+		}); err != nil {
+			return MultiPoint{}, MultiLineString{}, err
+		}
 	}
-	return NewMultiPoint(pts), NewMultiLineStringFromLineStrings(lss)
+	return NewMultiPoint(pts), NewMultiLineStringFromLineStrings(lss), nil
 }
 
 func intersectionOfMultiPointAndMultiPoint(mp1, mp2 MultiPoint) MultiPoint {
@@ -39,9 +45,10 @@ func intersectionOfMultiPointAndMultiPoint(mp1, mp2 MultiPoint) MultiPoint {
 	}
 	var pts []Point
 	for i := 0; i < mp2.NumPoints(); i++ {
-		xy, ok := mp2.PointN(i).XY()
+		pt := mp2.PointN(i)
+		xy, ok := pt.XY()
 		if ok && inMP1[xy] {
-			pts = append(pts, xy.AsPoint())
+			pts = append(pts, pt)
 		}
 	}
 	return NewMultiPoint(pts)