--- a/geom/alg_intersection.go
+++ b/geom/alg_intersection.go
@@ -3,24 +3,21 @@ package geom
 func intersectionOfIndexedLines(
 	lines1 indexedLines, lines2 indexedLines,
 ) (
-	MultiPoint, MultiLineString, error,
+	MultiPoint, MultiLineString,
 ) {
 	// TODO: Investigate potential speed up of swapping lines.
 	var lss []LineString
 	var pts []Point
 	seen := make(map[XY]bool)
 	for i := range lines1.lines {
-		if err := lines2.tree.RangeSearch(lines1.lines[i].uncheckedEnvelope().box(), func(j int) error {
+		lines2.tree.RangeSearch(lines1.lines[i].uncheckedEnvelope().box(), func(j int) error {
 			inter := lines1.lines[i].intersectLine(lines2.lines[j])
 			if inter.empty {
 				return nil
 			}
 			if inter.ptA == inter.ptB {
 				if xy := inter.ptA; !seen[xy] {
-					pt, err := xy.AsPoint()
-					if err != nil {
-						return wrap(err, "intersection between indexed lines is invalid")
-					}
+					pt := xy.asUncheckedPoint()
 					pts = append(pts, pt)
 					seen[xy] = true
 				}
@@ -28,11 +25,9 @@ func intersectionOfIndexedLines(
 				lss = append(lss, line{inter.ptA, inter.ptB}.asLineString())
 			}
 			return nil
-		}); err != nil {
-			return MultiPoint{}, MultiLineString{}, err
-		}
+		})
 	}
-	return NewMultiPoint(pts), NewMultiLineStringFromLineStrings(lss), nil
+	return NewMultiPoint(pts), NewMultiLineStringFromLineStrings(lss)
 }
 
 func intersectionOfMultiPointAndMultiPoint(mp1, mp2 MultiPoint) MultiPoint {