--- a/geom/type_envelope.go
+++ b/geom/type_envelope.go
@@ -10,22 +10,32 @@ import (
 // Envelope is an axis-aligned rectangle (also known as an Axis Aligned
 // Bounding Box or Minimum Bounding Rectangle). It usually represents a 2D area
 // with non-zero width and height, but can also represent degenerate cases
-// where the width or height (or both) are zero.
+// where the width or height (or both) are zero. Its bounds are validated so as
+// to not be NaN or +/- Infinity.
 type Envelope struct {
 	min XY
 	max XY
 }
 
 // NewEnvelope returns the smallest envelope that contains all provided points.
-func NewEnvelope(first XY, others ...XY) Envelope {
+// It returns an error if any of the XYs contains NaN or +/- Infinity
+// coordinates.
+func NewEnvelope(first XY, others ...XY) (Envelope, error) {
+	if err := first.validate(); err != nil {
+		return Envelope{}, err
+	}
 	env := Envelope{
 		min: first,
 		max: first,
 	}
-	for _, pt := range others {
-		env = env.ExtendToIncludePoint(pt)
+	for _, xy := range others {
+		var err error
+		env, err = env.ExtendToIncludeXY(xy)
+		if err != nil {
+			return Envelope{}, err
+		}
 	}
-	return env
+	return env, nil
 }
 
 // EnvelopeFromGeoms returns the smallest envelope that contains all points
@@ -93,12 +103,23 @@ func (e Envelope) Max() XY {
 	return e.max
 }
 
-// ExtendToIncludePoint returns the smallest envelope that contains all of the
-// points in this envelope along with the provided point.
-func (e Envelope) ExtendToIncludePoint(point XY) Envelope {
+// ExtendToIncludeXY returns the smallest envelope that contains all of the
+// points in this envelope along with the provided point. It gives an error if
+// the XY contains NaN or +/- Infinite coordinates.
+func (e Envelope) ExtendToIncludeXY(xy XY) (Envelope, error) {
+	if err := xy.validate(); err != nil {
+		return Envelope{}, err
+	}
+	return e.uncheckedExtend(xy), nil
+}
+
+// uncheckedExtend extends the envelope in the same manner as
+// ExtendToIncludeXY but doesn't validate the XY. It should only be used
+// when the XY doesn't come directly from user input.
+func (e Envelope) uncheckedExtend(xy XY) Envelope {
 	return Envelope{
-		min: XY{fastMin(e.min.X, point.X), fastMin(e.min.Y, point.Y)},
-		max: XY{fastMax(e.max.X, point.X), fastMax(e.max.Y, point.Y)},
+		min: XY{fastMin(e.min.X, xy.X), fastMin(e.min.Y, xy.Y)},
+		max: XY{fastMax(e.max.X, xy.X), fastMax(e.max.Y, xy.Y)},
 	}
 }
 
@@ -175,7 +196,7 @@ func (e Envelope) ExpandBy(x, y float64) (Envelope, bool) {
 	return env, true
 }
 
-// Distance calculates the stortest distance between this envelope and another
+// Distance calculates the shortest distance between this envelope and another
 // envelope. If the envelopes intersect with each other, then the returned
 // distance is 0.
 func (e Envelope) Distance(o Envelope) float64 {