--- a/geom/type_line_string.go
+++ b/geom/type_line_string.go
@@ -71,7 +71,7 @@ func (s LineString) StartPoint() Point {
 	if s.IsEmpty() {
 		return NewEmptyPoint(s.CoordinatesType())
 	}
-	return NewPoint(s.seq.Get(0))
+	return mustNewPoint(s.seq.Get(0))
 }
 
 // EndPoint gives the last point of the LineString. If the LineString is empty
@@ -80,7 +80,7 @@ func (s LineString) EndPoint() Point {
 	if s.IsEmpty() {
 		return NewEmptyPoint(s.CoordinatesType())
 	}
-	return NewPoint(s.seq.Get(s.seq.Length() - 1))
+	return mustNewPoint(s.seq.Get(s.seq.Length() - 1))
 }
 
 // AsText returns the WKT (Well Known Text) representation of this geometry.
@@ -325,7 +325,7 @@ func (s LineString) Centroid() Point {
 	if sumLength == 0 {
 		return NewEmptyPoint(DimXY)
 	}
-	return NewPointFromXY(sumXY.Scale(1.0 / sumLength))
+	return mustNewPointFromXY(sumXY.Scale(1.0 / sumLength))
 }
 
 func sumCentroidAndLengthOfLineString(s LineString) (sumXY XY, sumLength float64) {
@@ -390,7 +390,7 @@ func (s LineString) PointOnSurface() Point {
 	n := s.seq.Length()
 	nearest := newNearestPointAccumulator(s.Centroid())
 	for i := 1; i < n-1; i++ {
-		candidate := NewPointFromXY(s.seq.GetXY(i))
+		candidate := mustNewPointFromXY(s.seq.GetXY(i))
 		nearest.consider(candidate)
 	}
 	if !nearest.point.IsEmpty() {