… geom/accessor_test.go
… geom/ctor_options_test.go
… geom/dcel_interaction_points_test.go
… geom/perf_internal_test.go
… geom/perf_test.go
… geom/type_envelope_test.go
… geom/type_geometry_test.go
… geom/type_multi_point.go
… geom/type_multi_polygon.go
… geom/type_point.go
… geom/type_polygon.go
… geom/type_sequence.go
… geom/util_test.go
… geom/validation_test.go
… geom/wkb_parser.go
… geom/wkt_parser.go
… geom/xy.go
… geom/zero_value_test.go
… internal/cmprefimpl/cmpgeos/handle.go
… internal/cmprefimpl/cmpgeos/util_test.go