## Description

NaNs and Infs in geometries will cause unexpected issues with arithmetic operations. In most of the algorithms, `float64`s are treated as idealised [real numbers](https://en.wikipedia.org/wiki/Real_number). For the most part, this works ok for `float64` values that are not NaN or Inf. It breaks pretty badly for NaN and Inf values though.

### High Level Change Overview

- Geometries now have all of their coordinates checked for NaN and infinities when they are constructed. _Most_ geometries are constructed from other geometries, so changes were only needed to `Point` and `LineString`.
- Geometries are created based either on coordinates from existing geometries, or calculations based on existing geometries. When these new geometries are constructed, they don't always have their coordinates checked. This is because the NaN/Infinity checks only block user-supplied NaNs and Infinities. Error handling gets out of hand if we recheck NaN and infinities each time (e.g. when accessing the start point of LineString, it's not sensible to check the first coordinate to see if it's valid when constructing that point, since validation would have already occurred when the LineString was constructed). Similar logic holds for calculated geometries, e.g. centroid.
- Envelopes now have their coordinates checked. From an abstraction perspective, Envelopes are very similar to geometries, so should have similar NaN/Infinity behaviour. As a result, a few envelope methods now return errors.

### Breaking Changes

- The `NewEnvelope` function now returns `(Envelope, error)` rather than `error`.
- `Envelope`'s `ExtendToIncludeXY` method now returns `(Envelope, error)` rather than `error`.
- `Envelope`'s `ExpandBy` method was removed.
- The `NewLineString` function now returns an error if any of the X or Y values in its input sequence are NaN or an Infinity.
- The `NewPoint` function now returns `(Point, error)` rather than `Point`.
- `XY`'s `AsPoint` method now returns `(Point, error)` rather than `Point`.
- The `OmitEmpty` validation option now has implications when constructing `Point` and `MultiPoint` types.

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/396

## Benchmark Results

There are a few minor blips, but I think that has more to do with an inappropriate p-value cutoff.

<details>

<summary>Click to expand</summary>

```
COMPARISON
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.85ns ±50%    2.05ns ±49%    ~     (p=0.170 n=15+15)
LineEnvelope/1-4                                              1.06ns ± 3%    1.05ns ± 4%    ~     (p=0.667 n=14+14)
LineEnvelope/2-4                                              1.34ns ± 7%    1.31ns ± 3%    ~     (p=0.074 n=15+13)
LineEnvelope/3-4                                              1.18ns ± 4%    1.18ns ± 3%    ~     (p=0.838 n=15+14)
MarshalWKB/polygon/n=10-4                                      192ns ± 8%     194ns ±12%    ~     (p=0.674 n=13+15)
MarshalWKB/polygon/n=100-4                                     542ns ±19%     561ns ±14%    ~     (p=0.178 n=14+14)
MarshalWKB/polygon/n=1000-4                                   3.86µs ±37%    4.38µs ±68%    ~     (p=0.603 n=14+14)
MarshalWKB/polygon/n=10000-4                                  30.6µs ±29%    33.9µs ±51%    ~     (p=0.621 n=15+14)
UnmarshalWKB/polygon/n=10-4                                    316ns ±23%     313ns ±14%    ~     (p=0.901 n=13+15)
UnmarshalWKB/polygon/n=100-4                                   683ns ±31%     697ns ±20%    ~     (p=0.478 n=13+12)
UnmarshalWKB/polygon/n=1000-4                                 3.85µs ±29%    4.71µs ±83%    ~     (p=0.436 n=15+15)
UnmarshalWKB/polygon/n=10000-4                                40.3µs ±40%    39.1µs ±33%    ~     (p=0.880 n=14+15)
IntersectsLineStringWithLineString/n=10-4                     1.43µs ± 8%    1.49µs ±21%    ~     (p=0.340 n=12+13)
IntersectsLineStringWithLineString/n=100-4                    20.7µs ± 7%    21.3µs ± 8%    ~     (p=0.104 n=14+14)
IntersectsLineStringWithLineString/n=1000-4                    195µs ± 8%     198µs ± 5%    ~     (p=0.201 n=15+14)
IntersectsLineStringWithLineString/n=10000-4                  2.96ms ±16%    3.03ms ±13%    ~     (p=0.270 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20-4                      926ns ± 5%     937ns ± 7%    ~     (p=0.661 n=15+15)
IntersectsMultiPointWithMultiPoint/n=200-4                    10.4µs ± 4%    10.5µs ± 8%    ~     (p=0.659 n=15+13)
IntersectsMultiPointWithMultiPoint/n=2000-4                    103µs ± 7%     104µs ± 8%    ~     (p=0.813 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.18ms ± 9%    1.17ms ±11%    ~     (p=0.545 n=13+13)
PolygonSingleRingValidation/n=10-4                            2.42µs ±11%    2.38µs ± 6%    ~     (p=0.579 n=14+14)
PolygonSingleRingValidation/n=100-4                           32.5µs ± 6%    32.7µs ± 6%    ~     (p=0.793 n=14+13)
PolygonSingleRingValidation/n=1000-4                           386µs ± 7%     392µs ± 9%    ~     (p=0.486 n=15+15)
PolygonSingleRingValidation/n=10000-4                         4.95ms ±11%    4.97ms ±17%    ~     (p=0.583 n=14+13)
PolygonMultipleRingsValidation/n=4-4                          6.67µs ± 7%    6.87µs ± 8%  +2.95%  (p=0.027 n=14+15)
PolygonMultipleRingsValidation/n=36-4                         58.0µs ±12%    57.4µs ± 8%    ~     (p=0.964 n=15+13)
PolygonMultipleRingsValidation/n=400-4                         727µs ± 9%     744µs ± 5%  +2.43%  (p=0.038 n=13+14)
PolygonMultipleRingsValidation/n=4096-4                       8.64ms ± 8%    8.54ms ± 6%    ~     (p=0.595 n=15+15)
PolygonZigZagRingsValidation/n=10-4                           11.2µs ± 9%    11.2µs ±12%    ~     (p=1.000 n=13+13)
PolygonZigZagRingsValidation/n=100-4                           126µs ± 5%     129µs ± 7%    ~     (p=0.186 n=14+15)
PolygonZigZagRingsValidation/n=1000-4                         1.41ms ± 6%    1.46ms ±12%    ~     (p=0.233 n=15+15)
PolygonZigZagRingsValidation/n=10000-4                        18.5ms ± 4%    19.1ms ±11%  +3.38%  (p=0.006 n=13+14)
PolygonAnnulusValidation/n=10-4                               3.65µs ± 6%    3.81µs ±15%    ~     (p=0.166 n=13+15)
PolygonAnnulusValidation/n=100-4                              33.3µs ±12%    33.6µs ±14%    ~     (p=0.786 n=15+13)
PolygonAnnulusValidation/n=1000-4                              551µs ± 8%     533µs ± 5%    ~     (p=0.061 n=14+13)
PolygonAnnulusValidation/n=10000-4                            6.38ms ± 8%    6.35ms ± 5%    ~     (p=0.914 n=15+14)
MultipolygonValidation/n=1-4                                   399ns ±16%     391ns ±17%    ~     (p=0.259 n=14+13)
MultipolygonValidation/n=4-4                                   891ns ± 9%     916ns ±30%    ~     (p=0.277 n=12+15)
MultipolygonValidation/n=16-4                                 3.55µs ±11%    3.49µs ± 9%    ~     (p=0.427 n=14+14)
MultipolygonValidation/n=64-4                                 16.3µs ± 9%    16.7µs ±12%    ~     (p=0.302 n=13+14)
MultipolygonValidation/n=256-4                                 101µs ± 8%     102µs ± 8%    ~     (p=0.400 n=14+15)
MultipolygonValidation/n=1024-4                                479µs ± 6%     470µs ±10%    ~     (p=0.202 n=15+15)
MultiPolygonTwoCircles/n=10-4                                 3.58µs ± 8%    3.56µs ±10%    ~     (p=0.586 n=12+14)
MultiPolygonTwoCircles/n=100-4                                37.7µs ±12%    37.5µs ± 8%    ~     (p=0.920 n=13+13)
MultiPolygonTwoCircles/n=1000-4                                387µs ±10%     396µs ±10%    ~     (p=0.201 n=14+15)
MultiPolygonTwoCircles/n=10000-4                              5.60ms ±11%    5.73ms ±16%    ~     (p=0.603 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1-4                      4.81µs ± 7%    4.80µs ±11%    ~     (p=0.775 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-4                     37.6µs ± 9%    41.1µs ±39%    ~     (p=0.062 n=14+14)
MultiPolygonMultipleTouchingPoints/n=100-4                     438µs ± 4%     443µs ± 8%    ~     (p=0.456 n=12+15)
MultiPolygonMultipleTouchingPoints/n=1000-4                   5.03ms ± 7%    5.20ms ± 8%  +3.23%  (p=0.020 n=15+14)
WKTParsing/point-4                                            1.79µs ±22%    1.81µs ±14%    ~     (p=0.503 n=13+13)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           47.9µs ± 7%    47.0µs ± 4%    ~     (p=0.252 n=14+12)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            48.9µs ±14%    48.5µs ± 6%    ~     (p=1.000 n=14+13)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           742µs ± 3%     766µs ±13%    ~     (p=0.081 n=13+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            742µs ± 8%     759µs ± 6%    ~     (p=0.081 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     5.78µs ±22%    5.52µs ±16%    ~     (p=0.169 n=13+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      5.42µs ± 7%    5.43µs ± 9%    ~     (p=0.928 n=15+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    59.1µs ± 8%    58.9µs ±12%    ~     (p=0.591 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     58.6µs ±14%    60.3µs ±11%  +2.89%  (p=0.039 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=100-4                52.6µs ±12%    57.8µs ±32%    ~     (p=0.088 n=13+15)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                563µs ±16%     565µs ±17%    ~     (p=1.000 n=15+15)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-4                          39.1µs ± 6%    37.9µs ± 5%  -3.11%  (p=0.033 n=14+15)
IntersectionWithoutValidation/n=100-4                         74.7µs ± 5%    73.4µs ± 7%    ~     (p=0.077 n=14+14)
IntersectionWithoutValidation/n=1000-4                         334µs ± 9%     339µs ±11%    ~     (p=0.210 n=14+14)
IntersectionWithoutValidation/n=10000-4                       3.01ms ±15%    3.04ms ± 7%    ~     (p=0.496 n=13+15)
NoOp/n=10-4                                                   4.00µs ± 6%    3.93µs ± 3%    ~     (p=0.356 n=15+13)
NoOp/n=100-4                                                  11.9µs ±11%    12.3µs ±20%    ~     (p=0.539 n=15+15)
NoOp/n=1000-4                                                 83.9µs ± 7%    81.0µs ± 4%  -3.48%  (p=0.012 n=14+13)
NoOp/n=10000-4                                                 958µs ±12%     962µs ± 9%    ~     (p=0.715 n=15+14)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-4                               2.17µs ±14%    2.03µs ± 5%  -6.36%  (p=0.008 n=15+12)
LineStringIsSimpleCircle/n=100-4                              30.9µs ±11%    30.4µs ±15%    ~     (p=0.400 n=14+15)
LineStringIsSimpleCircle/n=1000-4                              371µs ± 5%     380µs ±10%    ~     (p=0.295 n=12+13)
LineStringIsSimpleCircle/n=10000-4                            4.79ms ±12%    4.98ms ±18%    ~     (p=0.400 n=14+15)
LineStringIsSimpleZigZag/10-4                                 1.90µs ±11%    1.86µs ± 7%    ~     (p=0.447 n=15+13)
LineStringIsSimpleZigZag/100-4                                31.2µs ±15%    30.4µs ± 6%    ~     (p=0.618 n=13+15)
LineStringIsSimpleZigZag/1000-4                                355µs ± 5%     363µs ± 5%  +2.08%  (p=0.020 n=14+15)
LineStringIsSimpleZigZag/10000-4                              4.82ms ± 3%    4.96ms ± 8%    ~     (p=0.070 n=15+14)
SetOperation/n=4/Go_Intersection-4                            42.7µs ±10%    43.9µs ±10%    ~     (p=0.137 n=14+14)
SetOperation/n=4/Go_Difference-4                              42.9µs ± 4%    43.2µs ± 9%    ~     (p=0.830 n=14+13)
SetOperation/n=4/Go_SymmetricDifference-4                     55.8µs ± 5%    56.3µs ± 7%    ~     (p=0.591 n=15+14)
SetOperation/n=4/Go_Union-4                                   44.7µs ± 3%    46.3µs ±14%    ~     (p=0.100 n=13+15)
SetOperation/n=4/GEOS_Intersection-4                          34.3µs ± 6%    34.2µs ± 1%    ~     (p=0.723 n=15+10)
SetOperation/n=4/GEOS_Difference-4                            35.6µs ± 6%    35.3µs ± 6%    ~     (p=0.591 n=15+14)
SetOperation/n=4/GEOS_SymmetricDifference-4                   49.7µs ± 7%    52.3µs ±10%  +5.14%  (p=0.003 n=14+15)
SetOperation/n=4/GEOS_Union-4                                 35.3µs ± 7%    35.6µs ± 4%    ~     (p=0.072 n=13+13)
SetOperation/n=8/Go_Intersection-4                            52.9µs ± 3%    55.2µs ±16%    ~     (p=0.118 n=12+14)
SetOperation/n=8/Go_Difference-4                              54.5µs ±10%    56.3µs ±24%    ~     (p=0.362 n=13+13)
SetOperation/n=8/Go_SymmetricDifference-4                     73.5µs ±18%    71.6µs ± 7%    ~     (p=0.683 n=15+13)
SetOperation/n=8/Go_Union-4                                   56.9µs ± 9%    57.5µs ± 7%    ~     (p=0.461 n=15+15)
SetOperation/n=8/GEOS_Intersection-4                          42.9µs ± 6%    43.8µs ± 7%    ~     (p=0.142 n=13+15)
SetOperation/n=8/GEOS_Difference-4                            42.6µs ± 4%    43.0µs ± 4%    ~     (p=0.254 n=13+15)
SetOperation/n=8/GEOS_SymmetricDifference-4                   59.9µs ± 7%    60.1µs ± 7%    ~     (p=0.747 n=15+14)
SetOperation/n=8/GEOS_Union-4                                 40.6µs ± 3%    41.5µs ± 5%  +2.23%  (p=0.037 n=12+15)
SetOperation/n=16/Go_Intersection-4                           76.7µs ±14%    75.3µs ± 4%    ~     (p=0.905 n=14+13)
SetOperation/n=16/Go_Difference-4                             80.5µs ± 7%    80.8µs ±16%    ~     (p=0.470 n=14+15)
SetOperation/n=16/Go_SymmetricDifference-4                     108µs ± 4%     110µs ±11%    ~     (p=0.980 n=12+14)
SetOperation/n=16/Go_Union-4                                  84.0µs ± 7%    84.8µs ± 7%    ~     (p=0.377 n=14+15)
SetOperation/n=16/GEOS_Intersection-4                         47.9µs ± 6%    47.5µs ± 5%    ~     (p=0.539 n=15+15)
SetOperation/n=16/GEOS_Difference-4                           50.8µs ± 8%    49.4µs ± 4%    ~     (p=0.051 n=15+14)
SetOperation/n=16/GEOS_SymmetricDifference-4                  76.8µs ± 4%    79.8µs ± 7%  +3.84%  (p=0.044 n=14+14)
SetOperation/n=16/GEOS_Union-4                                51.1µs ± 9%    51.3µs ± 8%    ~     (p=0.667 n=14+14)
SetOperation/n=32/Go_Intersection-4                            129µs ± 6%     130µs ± 6%    ~     (p=0.618 n=13+15)
SetOperation/n=32/Go_Difference-4                              136µs ± 6%     135µs ± 8%    ~     (p=0.561 n=14+15)
SetOperation/n=32/Go_SymmetricDifference-4                     183µs ± 6%     180µs ± 8%    ~     (p=0.451 n=14+15)
SetOperation/n=32/Go_Union-4                                   141µs ± 7%     140µs ± 6%    ~     (p=0.685 n=13+14)
SetOperation/n=32/GEOS_Intersection-4                         59.4µs ± 7%    58.5µs ± 4%    ~     (p=0.390 n=13+13)
SetOperation/n=32/GEOS_Difference-4                           63.3µs ± 9%    61.1µs ± 4%  -3.49%  (p=0.019 n=15+13)
SetOperation/n=32/GEOS_SymmetricDifference-4                   111µs ±11%     109µs ± 6%    ~     (p=0.607 n=14+13)
SetOperation/n=32/GEOS_Union-4                                62.6µs ± 5%    62.5µs ± 4%    ~     (p=0.983 n=15+14)
SetOperation/n=64/Go_Intersection-4                            222µs ±12%     216µs ± 9%    ~     (p=0.354 n=14+15)
SetOperation/n=64/Go_Difference-4                              233µs ± 8%     230µs ± 6%    ~     (p=0.482 n=14+14)
SetOperation/n=64/Go_SymmetricDifference-4                     316µs ± 7%     316µs ± 9%    ~     (p=0.964 n=15+13)
SetOperation/n=64/Go_Union-4                                   248µs ± 7%     243µs ±10%    ~     (p=0.231 n=14+12)
SetOperation/n=64/GEOS_Intersection-4                         78.7µs ± 4%    79.7µs ± 4%    ~     (p=0.294 n=15+13)
SetOperation/n=64/GEOS_Difference-4                           95.4µs ± 7%    96.8µs ± 6%    ~     (p=0.178 n=14+14)
SetOperation/n=64/GEOS_SymmetricDifference-4                   183µs ± 7%     186µs ± 7%    ~     (p=0.302 n=13+14)
SetOperation/n=64/GEOS_Union-4                                 102µs ±12%     104µs ± 8%    ~     (p=0.063 n=14+15)
SetOperation/n=128/Go_Intersection-4                           400µs ± 5%     414µs ±11%    ~     (p=0.146 n=14+15)
SetOperation/n=128/Go_Difference-4                             425µs ± 7%     444µs ±13%  +4.36%  (p=0.045 n=15+15)
SetOperation/n=128/Go_SymmetricDifference-4                    561µs ± 7%     563µs ± 5%    ~     (p=0.525 n=13+15)
SetOperation/n=128/Go_Union-4                                  447µs ±12%     449µs ±11%    ~     (p=0.511 n=14+14)
SetOperation/n=128/GEOS_Intersection-4                         131µs ±13%     132µs ± 9%    ~     (p=0.652 n=14+15)
SetOperation/n=128/GEOS_Difference-4                           149µs ±11%     152µs ±10%    ~     (p=0.105 n=13+14)
SetOperation/n=128/GEOS_SymmetricDifference-4                  299µs ± 6%     315µs ±20%    ~     (p=0.123 n=15+14)
SetOperation/n=128/GEOS_Union-4                                156µs ± 6%     158µs ±10%    ~     (p=0.856 n=13+15)
SetOperation/n=256/Go_Intersection-4                           731µs ± 7%     736µs ± 9%    ~     (p=0.683 n=15+14)
SetOperation/n=256/Go_Difference-4                             787µs ± 6%     792µs ± 8%    ~     (p=0.839 n=14+14)
SetOperation/n=256/Go_SymmetricDifference-4                   1.10ms ± 8%    1.12ms ±12%    ~     (p=0.217 n=12+15)
SetOperation/n=256/Go_Union-4                                  818µs ± 6%     826µs ±12%    ~     (p=0.943 n=15+12)
SetOperation/n=256/GEOS_Intersection-4                         205µs ±12%     199µs ± 5%    ~     (p=0.914 n=15+14)
SetOperation/n=256/GEOS_Difference-4                           260µs ± 7%     259µs ± 6%    ~     (p=0.650 n=13+14)
SetOperation/n=256/GEOS_SymmetricDifference-4                  556µs ± 5%     547µs ± 9%    ~     (p=0.110 n=12+13)
SetOperation/n=256/GEOS_Union-4                                276µs ± 6%     277µs ± 4%    ~     (p=0.821 n=15+13)
SetOperation/n=512/Go_Intersection-4                          1.53ms ±16%    1.47ms ±10%    ~     (p=0.201 n=14+15)
SetOperation/n=512/Go_Difference-4                            1.60ms ±11%    1.55ms ±15%  -3.33%  (p=0.037 n=15+14)
SetOperation/n=512/Go_SymmetricDifference-4                   2.15ms ± 6%    2.13ms ± 7%    ~     (p=0.496 n=13+15)
SetOperation/n=512/Go_Union-4                                 1.61ms ± 7%    1.65ms ±18%    ~     (p=0.874 n=14+14)
SetOperation/n=512/GEOS_Intersection-4                         358µs ± 6%     357µs ± 5%    ~     (p=0.983 n=14+15)
SetOperation/n=512/GEOS_Difference-4                           437µs ±11%     428µs ± 5%    ~     (p=0.983 n=14+15)
SetOperation/n=512/GEOS_SymmetricDifference-4                 1.06ms ±14%    1.05ms ±15%    ~     (p=0.847 n=15+14)
SetOperation/n=512/GEOS_Union-4                                479µs ± 7%     470µs ± 4%    ~     (p=0.235 n=15+13)
SetOperation/n=1024/Go_Intersection-4                         2.89ms ±10%    2.85ms ±10%    ~     (p=0.400 n=14+15)
SetOperation/n=1024/Go_Difference-4                           3.13ms ±11%    3.08ms ± 8%    ~     (p=0.595 n=14+12)
SetOperation/n=1024/Go_SymmetricDifference-4                  4.27ms ± 5%    4.34ms ± 5%    ~     (p=0.085 n=13+14)
SetOperation/n=1024/Go_Union-4                                3.25ms ± 9%    3.17ms ± 7%    ~     (p=0.137 n=14+14)
SetOperation/n=1024/GEOS_Intersection-4                        655µs ± 8%     659µs ± 9%    ~     (p=0.715 n=14+15)
…