@albertteoh 2021-09-08
LGTM!

I think you meant the following in the description: "The `NewEnvelope` function now returns `(Envelope, error)` rather than `Envelope`."

Same with `ExtendToIncludeXY`.

Also, I also think you meant: "The `OmitInvalid` validation option now has implications when constructing `Point` and `MultiPoint` types."

Instead of: "The `OmitEmpty` validation option now has implications when constructing `Point` and `MultiPoint` types."