@peterstace 2021-09-08
> I think you meant the following in the description

Thanks! There were a bunch of other errors and improvements to the release notes as well when I looked back at them. I've added a bunch of changes in a PR here: https://github.com/peterstace/simplefeatures/pull/415

I think maybe in the future we should be adding release notes in an "unreleased" section in the CHANGELOG during regular PRs. That way, the release process is just updating the CHANGELOG headers rather than having to write release notes from scratch.